refactor(cmd/up): move agent and configure files into up subpackage
After rebasing onto main (which added cmd/up_agent.go and
cmd/up_configure.go in package cmd via PR #228), move these files into
cmd/up/ as package up so UpCmd methods remain accessible.
