+++ cmd/up/configure.go
@@ -1,4 +1,4 @@
-package cmd
+package up
 
 import (
 	"context"
